Begin migration from MySQL to PostgreSQL. Configure package to use workspaces
