+++ README.md
@@ -2,18 +2,22 @@
 
 A web based application to provide point of sales services for small businesses.
 
-Currently working on implementing data layer for a few categories.
+Currently working on implementing data layer for API
 
-Upcoming work includes:
-1. API Logging
-2. Making Data Access for: Stock Category, Tax, Product, Supplier, Customer, Transaction
-3. Basic react page for transaction
+Road Map:
+1. Convert app to use PostgreSQL
+1. Making Data Access for: Stock Category, Tax, Product, Supplier, Customer, Transaction
+1. Explore and implement JSON Web Token for authentication 
+1. Create prototype react page for transactions
+1. Create complex queries
 
 Client Side Tech Stack:
 - React
+- TypeScript
 
 API Tech Stack:
 - ExpressJS
-- TypeScript
-- MySQL
 - Bunyan (Logger)
+- TypeScript
+- <s>MySQL</s> PostgreSQL
+- TypeORM
