+++ package.json
@@ -1,8 +1,21 @@
 {
   "name": "retail-point-of-sale",
   "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "private": true,
+  "workspaces": [
+    "packages\\server",
+    "packages\\client"
+  ],
+  "dependencies": {
+    "typescript": "^4.9.4",
+    "bunyan": "^1.8.15"
+  },
+  "devDependencies": {
+    "@types/node": "^18.11.15",
+    "@types/bunyan": "^1.8.8",
+    "nodemon": "^2.0.20",
+    "ts-node": "^10.9.1"
+  },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "npx ts-node src/Server/server.ts",
@@ -12,25 +25,6 @@
     "type": "git",
     "url": "git+https://github.com/calmacleod/Retail_PointOfSale.git"
   },
-  "keywords": [],
-  "author": "",
   "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/calmacleod/Retail_PointOfSale/issues"
-  },
-  "homepage": "https://github.com/calmacleod/Retail_PointOfSale#readme",
-  "dependencies": {
-    "@types/bunyan": "^1.8.8",
-    "bunyan": "^1.8.15",
-    "express": "^4.18.1",
-    "express-bunyan-logger": "^1.3.3",
-    "mysql2": "^2.3.3"
-  },
-  "devDependencies": {
-    "@types/mysql2": "github:types/mysql2",
-    "@types/node": "^18.11.15",
-    "nodemon": "^2.0.20",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.9.4"
-  }
+  "homepage": "https://github.com/calmacleod/Retail_PointOfSale#readme"
 }
