+++ packages/client/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "client",
+  "version": "1.0.0",
+  "description": "React client for Retail Point of Sale Application",
+  "main": "index.js",
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/calmacleod/Retail_PointOfSale.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/calmacleod/Retail_PointOfSale/issues"
+  },
+  "homepage": "https://github.com/calmacleod/Retail_PointOfSale#readme"
+}
