+++ packages/server/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "server",
+  "version": "1.0.0",
+  "description": "Web server for Retail Point of Sale application",
+  "main": "index.ts",
+  "devDependencies": {},
+  "dependencies": {
+    "express": "^4.18.1"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "npx ts-node src/index.ts",
+    "start:dev": "npx nodemon"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/calmacleod/Retail_PointOfSale.git"
+  },
+  "license": "ISC",
+  "homepage": "https://github.com/calmacleod/Retail_PointOfSale#readme"
+}
