+++ packages/server/src/credentials.json
@@ -1,0 +1,4 @@
+{
+    "mysql_user" : "root",
+    "mysql_pass" : "Ilovemax12!"
+}
