+++ packages/server/src/pages/product_entry.html
