--- src/Server/Data_Access/ModelDBs/StockCategoryDB.ts
@@ -1,6 +1,0 @@
-import { IStockCategory } from "../Models/StockCategoryModel";
-
-export interface StockCategoryDB {
-    findAllCategories() : Promise<IStockCategory[]>;
-    findByID(id:number) : Promise<IStockCategory | undefined>;
-}
