--- src/Server/Data_Access/ModelDBs/TaxDB.ts
@@ -1,6 +1,0 @@
-import { ITax } from "../Models/TaxModel";
-
-export interface TaxDB {
-    findAllCodes() : Promise<ITax[]>;
-    findByCode(code:number) : Promise<ITax | undefined>;
-}
