--- src/Server/Data_Access/Models/StockCategoryModel.ts
@@ -1,6 +1,0 @@
-import { RowDataPacket } from "mysql2";
-
-export interface IStockCategory extends RowDataPacket {
-    tax_id: number;
-    stock_name: string;
-};
