--- src/Server/Data_Access/Models/TaxModel.ts
@@ -1,7 +1,0 @@
-import { RowDataPacket } from "mysql2";
-
-export interface ITax extends RowDataPacket {
-    tax_code: number;
-    tax_name: string;
-    tax_rate: number;
-};
