--- src/Server/Data_Access/Models/product_model.ts
@@ -1,13 +1,0 @@
-export type ProductModel = {
-    stock_code: string;
-    product_name: string;
-    selling_price: number;
-    purchase_price: number;
-    stock_level: number;
-    date_added: string;
-    stock_category: number;
-    reorder_level: number;
-    unit_cost: number;
-    items_unit: number;
-    supplier_id: number;
-};
