--- src/Server/Data_Access/Queries/product
@@ -1,12 +1,0 @@
-const data = require('../data')
-
-function get_product(product_id) {
-    var pool = data.myData.getPool();
-
-    var SQL_Select_Product = 'SELECT * FROM stock WHERE stock_code = ?'
-    
-    
-
-
-    data.myData.getPool().query()
-}
