--- src/Server/Routes/product.js
@@ -1,2 +1,0 @@
-const express = require('express');
-const router = express.Router();
